Rewrote the Customer Router and View to use a bind/trigger model
More efficient and DRY than the previous method. It's not perfectly DRY
but it means that I can effectively trigger new refreshes for content
when pages have been updated.

The next update should see a few new events that trigger when adding
items to a customer so that the extras lists are updated with new info