JS now modularised into git submodules and renamed minify plugin
Backbone boilerplate is now a submodule

The dux app is now minified with the sma plugin and wrapped in an
anonymous function and added to the window object to avoid namespace
pollution.

View templates are updated to refer to the Sma plugin helper and
components now refer to the Sma plugin component.