Improved Search
Tidied up the code so that it relies more upon the collection results
rather than input to show/hide the search results view.

Tied events to call collection.reset and collection.fetch.

Added the router to the Search view

Also, I removed the global event catcher that was catching customer URIs
globally and made it restricted to [role=main]