Views and Routers changed with new cbb update
Backbone, Underscore and my connrsBackboneBoilerplate have been updated
meaning that I wanted to refactor some of this. Additionally, I've
created individual list views for several models that extend the core
list view as I wanted to tidy the code for reuse and reduce some of the
overly verbose creation of views inside the customers view itself (as it
creates several sublist views) because it was fairly unwieldy.

Several files had their indent level altered as it was indented back
before I refactored the classes inside individual file enclosures which
are now package enclosures